% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_basic_neutral}
\alias{data_basic_neutral}
\title{data_basic_neutral}
\format{
A data frame with 319 rows and 26 variables:
\describe{
\item{\code{cooperation}}{character Reputations based on the willingness, experience, capacities, or the likelihood an individual will (or will not) intentionally work together with or assist another individual or group or contribute to the efforts of another individual or group in order to achieve a joint goal or outcome.}
\item{\code{prosociality}}{character Reputations based on an individual's willingness, experience, capacities or the likelihood one will (or will not) invest in group welfare or act in group-altruistic ways (i.e., in ways that are costly to the self). This definition includes emotions and dispositional characteristics that indicate a desire for sociality, such as friendliness.}
\item{\code{honesty}}{character Reputations for the quality of being honest or not lying or intentionally deceiving others (as well as those dispossessing this trait – those who transmit non-truthful information, especially with an intent to deceive).}
\item{\code{cultural_conformity}}{character Reputations based on excelling at, consistently demonstrating, or embodying skills or qualities preferentially valued within a particular cultural or social context. Includes culturally appropriate, preferred, and prescribed behaviours (e.g. willingness or ability to follow norms) (includes individuals who dispossess this trait).}
\item{\code{industriousness}}{character Reputations based on productive outputs from activities such as subsistence and other economic efforts, domestic work, recreational accomplishments, etc., including past and current productivity (includes individuals who dispossess this trait).}
\item{\code{coercive_ability}}{character Reputations based on others fear, threats, or intimidation rooted in power-imbalances and differences in physical formidability, and when these actions are associated with greater access to contested resources and/or social influence and control (includes individuals who dispossess this trait).}
\item{\code{agressiveness}}{character Reputations for a propensity to act aggressively and agonistically, though which may not include control or influence (includes individuals who dispossess this trait).}
\item{\code{social_status}}{character Reputations for having a high rank within social hierarchies, prestige, or esteem (or their opposite). Includes “global” social status, conceived broadly and likely dependent on multiple criteria. Separate from “coercive ability” reputations.}
\item{\code{parental_care}}{character Reputations for being a good parent, successfully raising many/quality children, and/or investing a lot of time, resources, and energy in raising children. (includes individuals who dispossess this trait).}
\item{\code{sociosexuality}}{character Reputations for having many sexual partners and/or willingness to engage in sexual activity or develop multiple sexual/romantic relationships (e.g., high sociosexuality) (includes individuals who dispossess this trait, e.g., low sociosexuality).}
\item{\code{sexual_fidelity}}{character Reputations for fidelity, sexual exclusivity, and maintaining physical commitment within a marriage or sexual relationship (includes individuals who dispossess this trait).}
\item{\code{somatic_capital}}{character Reputations for beneficial and positive physiological traits including health, strength, agility and dexterity (includes individuals who dispossess this trait).}
\item{\code{neural_capital}}{character Reputations for cognitive and neural-based dimensions of somatic or embodied capital, including reputations for generalized or specialized intelligence, special knowledge, cognitive abilities, and decision-making capacities (includes individuals who dispossess this trait).}
\item{\code{social_capital}}{character Reputations for having a large and/or strong social support network including having many social partners, kin or allies and having social partners, kin, or allies who are of high quality (includes individuals who dispossess this trait).}
\item{\code{material_capital}}{character Reputations for having high levels of material wealth, including all forms of material wealth and currencies, such as money, territory, prestige goods, property, having food reserves, and other material resources (includes individuals who dispossess this trait).}
\item{\code{supernatural_ability}}{character Reputations for having supernatural abilities and capacities beyond “normal” abilities, including capacities rooted in physical displays with an important supernatural component (such as trance) or other non-physical abilities (augury, supernatural abilities in ritual, magic, etc.) This can include healing related to supernatural agency(ies) (includes individuals who dispossess these traits).}
\item{\code{medicine}}{character Reputations for having (or not having) knowledge and ability to heal the self and others through naturalistic means.}
\item{\code{teaching}}{character Reputations for being a good teacher including all forms of teaching, defined as an informed individual modifying their behaviour for the learning benefit of a naive individual. Includes capacities to train and transmit specialized or generalized information. (includes individuals who dispossess this trait).}
\item{\code{oration}}{character Reputations for having (or not having) a superior ability to use language or other forms of vocalization in public and private contexts (includes individuals who dispossess this trait).}
\item{\code{bravery}}{character Reputations for the ability or disposition to voluntarily act when facing personal risk (in terms of agony, pain, loss, danger, uncertainty, or intimidation) in pursuit of a worthy goal, often in the face of fear (includes individuals who dispossess this trait).}
\item{\code{text_ID}}{character Internally assigned unique identifer for each text record.}
\item{\code{document_ID}}{character The HRAF document ID for each source document.}
\item{\code{culture_ID}}{character The HRAF Old World Cultures (OWC) unique identifier for each culture.}
\item{\code{sex}}{character Gender coding. All neutral here.}
}
}
\source{
\url{https://ehrafworldcultures.yale.edu/}
}
\usage{
data_basic_neutral
}
\description{
Basic researcher-coded variables for gender-neutral evidence only, raw ethnographic text and identifiers. Reputation domains coded as 1 for evidence for gender-neutral reputations and 0 for no evidence for gender-neutral reputations.
}
\keyword{datasets}
