% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_culture}
\alias{data_culture}
\title{data_culture}
\format{
A data frame with 153 rows and 10 variables:
\describe{
\item{\code{culture_ID}}{character The HRAF Old World Cultures (OWC) unique identifier for each culture.}
\item{\code{n}}{integer The number of documents from the eHRAF providing ethnographic texts for each culture.}
\item{\code{Region}}{character The eHRAF region classification for the culture's continental geographic location.}
\item{\code{Subregion}}{character The eHRAF sub-region classification for the culture's specific geographic location.}
\item{\code{Subsistence.Type}}{character The culture's eHRAF subsistence type classification.}
\item{\code{EA_ID}}{character The culture's Ethnographic Atlas identification number.}
\item{\code{SCCS}}{integer The culture's Standard Cross-Cultural Sample identification number.}
\item{\code{longitude}}{double Approximate longitude of culture geographic location.}
\item{\code{latitude}}{double Approximate latitude of culture geographic location.}
\item{\code{Name}}{character Culture name (eHRAF culture name).}
\item{\code{language_family}}{character Culture language family (retrieved from The Ethnographic Atlas or other sources).}
\item{\code{N_docs}}{double Number of total available eHRAF documents for culture.}
\item{\code{N_pages}}{double Number of total pages in all eHRAF documents for culture}
}
}
\source{
\url{https://ehrafworldcultures.yale.edu/}
}
\usage{
data_culture
}
\description{
Culture-level metadata.
}
\keyword{datasets}
