% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reputation_dtm}
\alias{reputation_dtm}
\title{reputation_dtm}
\format{
A data frame with 319 rows and 8770 variables:
\describe{
\item{\code{text_ID}}{character Internally assigned unique identifer for each text record.}
}
}
\source{
\url{https://ehrafworldcultures.yale.edu/}
}
\usage{
reputation_dtm
}
\description{
A document-term matrix formulated from the raw ethnographic texts (i.e., "Text" in data_raw). Each row corresponds to the ethnographic texts. Each column represents each unique term in the entire corpus of ethnographic texts. Cells indicate the frequency each term appears in the ethnographic text.
}
\keyword{datasets}
