% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reputation_words}
\alias{reputation_words}
\title{reputation_words}
\format{
A data frame with 45007 rows and 3 variables:
\describe{
\item{\code{text_ID}}{character Internally assigned unique identifer for each text record.}
\item{\code{word}}{character The original term.}
\item{\code{lemma}}{character The terms associated token.}
}
}
\source{
\url{https://ehrafworldcultures.yale.edu/}
}
\usage{
reputation_words
}
\description{
Database of all words and their associated terms (i.e., tokens)
}
\keyword{datasets}
